+++ Meraki.Api/Data/NetworkBluetoothSettingsTransmitAdvertised.cs
@@ -1,0 +1,15 @@
+namespace Meraki.Api.Data;
+
+/// <summary>
+/// Advertised transmit settings for Bluetooth beacon
+/// </summary>
+[DataContract]
+public class NetworkBluetoothSettingsTransmitAdvertised
+{
+	/// <summary>
+	/// Advertised transmit power
+	/// </summary>
+	[ApiAccess(ApiAccess.ReadUpdate)]
+	[DataMember(Name = "power")]
+	public int? Power { get; set; }
+}